Create README.md
Web scrapped IPL Scoresheets 

This is an introductory project focusing on the core concepts of web scraping. The aim of the project is to compile the attacking statistics of every player of all teams that took part in the sports competition: IPL 2021 and store the resultant data i the form of excel sheet for easy data analysis. The data has been acquired from the ESPN Cricket database's coverage of the IPL 2021 competition.

To run the program, I recommend using the VSCode application and simply typing "node './main.js'" in the integrated terminal allocated with the target directory destination. A new folder named 'IPL' is created thereafter. The folder is further composed of subfolders pertaining to each team that participated in the competition and inside the latter folders, player stats of the team are stored in the form of excel sheets.